backend/requirements.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.4.4

# AI Integration
openai==1.47.1
anthropic==0.40.0

# File Handling
Pillow==10.4.0

# Background Tasks (Optional)
celery==5.3.4
redis==5.0.1

# Development
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
